feat: changed typetest and result into children of Home
