+++ front/src/components/Result.tsx
